Add "Go To Definition"
When in a workspace, enable "Go To Definition". The extension assumes
that the first entity it finds with the target name is the desired
entity. This should hopefully, usually be correct. Because the extension
now makes use of typescript, tsconfig is added, and the package file
specifies an entry point into the code.

The latest version of the FSH grammar at time of writing is used.