Add test suite
A suite of tests for two of the extension's functions are added. The two
tested functions are getDefinitionLocation and getTargetName. The
runTest.ts and suite/index.ts are copied from the extension development
starter script. More information about that script is at
https://code.visualstudio.com/api/get-started/your-first-extension.

The mocha test library is used because that is what the generated code
expects. Other libraries can be used, but would require additional
effort. I tried to make it work with jest, but I was unable to make that
work. So, mocha it is.

To run the tests, run `npm run test` from the command line. Note that
this will usually fail if you have VSCode open, which is less than
convenient, but I don't know of a good way to dodge that problem.