Add "Go To Definition" (#11)
* Add "Go To Definition"

When in a workspace, enable "Go To Definition". The extension assumes
that the first entity it finds with the target name is the desired
entity. This should hopefully, usually be correct. Because the extension
now makes use of typescript, tsconfig is added, and the package file
specifies an entry point into the code.

The latest version of the FSH grammar at time of writing is used.

* Use custom word detection regex

The default word detection regex does not permit some characters that
may appear in an entity name. Use a custom regex that more closely
matches the SEQUENCE token rule.

Trim RuleSet name before comparing to target string when trying to find
a definition, since the RULESET_REFERENCE token will usually have a
leading space.

* Add eslint and prettier

Now that there is code and not just configuration in this extension, add
eslint and prettier so that it is formatted the same as other FSH
projects. Add script definitions to perform linting and prettification.

* Add grammar updating information to readme

* Do not match parentheses when reading entity name

If parentheses are included in the entity name, the extension will be
unable to find the definition location of an entity used in a reference,
since those are surrounded by parentheses and are not required to have
space between the entity name and the parentheses.

* Add Go to Definition to readme

* Remove local path from launch configuration

* Move NPM tasks in README to H2 level

* Add test suite

A suite of tests for two of the extension's functions are added. The two
tested functions are getDefinitionLocation and getTargetName. The
runTest.ts and suite/index.ts are copied from the extension development
starter script. More information about that script is at
https://code.visualstudio.com/api/get-started/your-first-extension.

The mocha test library is used because that is what the generated code
expects. Other libraries can be used, but would require additional
effort. I tried to make it work with jest, but I was unable to make that
work. So, mocha it is.

To run the tests, run `npm run test` from the command line. Note that
this will usually fail if you have VSCode open, which is less than
convenient, but I don't know of a good way to dodge that problem.

* Add CI cleaning and testing configuration

This configuration is copied from SUSHI, and runs the lint, prettier,
and test tasks.

* Use xvfb-run for testing on Linux environments

Based on documentation from Microsoft, xvfb is required to run VSCode
successfully as part of a GitHub Action. See
https://code.visualstudio.com/api/working-with-extensions/continuous-integration#github-actions
for more information.

* Clean up test output, ignore .vscode-test styles

* Detect code system aliases used in concepts

By adding # to the list of prohibited characters when getting the target
word for Go to Definition, it is possible to find the definition of an
alias when it is used as the system for a concept.

* Add gif for go to definition

Co-authored-by: Chris Moesel <cmoesel@mitre.org>