Implement documentation opening
This commit contains the actual implementation of the functions to
support the "Open Documentation" feature. FSH entity names have their
documentation URLs specified. Other names are assumed to be FHIR
entities with accompanying URLs. A HEAD request is performed to check
for the existence of the URL. If it succeeds, the extension will attempt
to open that URL in a browser. If it fails, a failure message is shown.

The axios library is added to make HTTP requests, and the nock library
is added to mock HTTP requests in the test suite.