Cache parse results in FshDefinitionProvider
FshDefinitionProvider maintains information gathered from previous
parses in order to speed up definition retrieval. When fsh files change
in the workspace, redo the parse on those files and update the cache.
When a definition is requested, reparse files with unsaved changes, then
return location information from the cache. A name may have multiple
locations associated with it. If this happens, return all the locations.

Because FshDefinitionProvider now has a lot more code, it is in its own
file.