Handle filesystem events on directories
When a directory is moved or deleted, a filesystem event occurs for the
directory, not its contents. Therefore, when receiving a filesystem
event, collect all applicable FSH files based on the path for the event.
Moving a directory inside a workspace will fire two events: a create
event at the move target, and a change event at the parent of the move
target. In order to avoid parsing files twice, store the hash of a file
when parsing it. If the incoming hash matches the last known hash, do
not parse the file again. This hash-comparison technique is also used
when checking files with unsaved changes.