Improve performance of "Go to definition" (#15)
* Definition provider returns list of Locations

More than one FSH entity may be defined with the same name. Return a
list of all possible locations when searching for a definition.

* Cache parse results in FshDefinitionProvider

FshDefinitionProvider maintains information gathered from previous
parses in order to speed up definition retrieval. When fsh files change
in the workspace, redo the parse on those files and update the cache.
When a definition is requested, reparse files with unsaved changes, then
return location information from the cache. A name may have multiple
locations associated with it. If this happens, return all the locations.

Because FshDefinitionProvider now has a lot more code, it is in its own
file.

* Handle filesystem events on directories

When a directory is moved or deleted, a filesystem event occurs for the
directory, not its contents. Therefore, when receiving a filesystem
event, collect all applicable FSH files based on the path for the event.
Moving a directory inside a workspace will fire two events: a create
event at the move target, and a change event at the parent of the move
target. In order to avoid parsing files twice, store the hash of a file
when parsing it. If the incoming hash matches the last known hash, do
not parse the file again. This hash-comparison technique is also used
when checking files with unsaved changes.