Provide contextual autocomplete for Parent, InstanceOf, and obeys
When specifying an entity's Parent, autocomplete will show the names of
FSH entities in the workspace that could serve as the parent. The types
of entities listed depends on what kind of entity is having its parent specified.

When specifying an Instance's InstanceOf, autocomplete will show the
names of FSH Profiles, Resources, and Extensions in the workspace.

When writing an obeys rule, autocomplete will show the names of FSH
Invariants in the workspace.