Add autocomplete items from FHIR package
By configuring a path to the FHIR cache, the extension reads the index
file from the FHIR R4 package. This index file provides information
about FHIR Resources, Extensions, CodeSystems, and ValueSets. When
listing these items as autocomplete results, prefix the name with "FHIR"
to indicate that they come from the package and not a local definition.