Add Profiles from package index files
A Profile will appear in an index file as a StructureDefinition whose
type and id are different. Most packages keep examples in a separate
folder, but some packages, such as FHIR R4 4.0.1, keep them together.
They are differentiated by the presence of a version attribute on
profiles, but not on examples.