Make autocomplete items from FHIR definition files
The index file in packages is no longer used. Instead, the FHIR
definition files are read in order to add autocomplete items. This
allows the name of the resource to be used when available. It also means
that packages without index files are supported.

Logicals are now detected and added to the set of autocomplete items for
FHIR packages.