Remove FHIR cache configuration
The path to the FHIR cache is no longer specified in configuration.
Instead, the path is assumed to be the same one that SUSHI uses.

When making completion items from a package, check if they are in the
cache. If so, used the cached items. If not, create new items and add
them to the cache.