Collect additional metadata when storing entities
When storing information about a Profile, parent should be available.
When storing information about an Extension, parent may be available,
but the default is frequently used. When storing information about an
Instance, instanceOf should be available. These fields are useful as
part of the rule path completion process, because they can lead to a
base definition.