Get element paths for a named entity
When loading FHIR dependencies, store the element paths from the
snapshot. The paths are stored with the existing entity information.
Then, when looking up element paths, traverse through local definitions
based on "Parent" or "InstanceOf" keyword values until a value that is
not present in the local definitions is reached. This value is hopefully
present in the dependency information. If the value is null, it may be
possible to use the default parent for the entity type.