Store additional element path information
When storing the element information for a FHIR entity, store the path
name and types. Additionally, use a tree structure to store path
information, rather than keeping all element paths as a flat list. When
reading in predefined resources, when a profile is found, rather than
reading its elements, use its type field in order to store the entity
being profiled. Then, refer to that entity when searching for paths for
that profile.