Create ElementInfo for each type allowed by a choice element
When a choice element is found in a FHIR entity, add the choice
element's info as well as an ElementInfo for each type available on the
choice element. These elements are siblings. Note that if the type of
the choice element is constrained by an OnlyRule, the extension does not
detect this, and will continue to offer the excluded types as completion
items.