Expand leaf elements to provide completion items
When attempting to complete a path, if the path defines an element that
has no known children but has a known type, use the type to create the
children of that element. When creating children for an element with
more than one type, only include new elements if they are compatible
with all types.

Fix bug in FshCompletionProvider.getElementPathInformation where the
name of a choice element would be incomplete due to overzealous slice
name removal.