Support URL references for completion items
When creating completion items from dependency definitions, create an
element for the URL if it is available. When creating completion items
from an Alias, store the value of the Alias. Together, this makes it
possible to use an Alias as a parent and get the correct elements for
rule path completion.

Make most properties in NameInfo optional, since many of them are
applicable only to specific entity types.