Provide autocompletion for paths when writing rules (#36)
* Collect additional metadata when storing entities

When storing information about a Profile, parent should be available.
When storing information about an Extension, parent may be available,
but the default is frequently used. When storing information about an
Instance, instanceOf should be available. These fields are useful as
part of the rule path completion process, because they can lead to a
base definition.

* Get element paths for a named entity

When loading FHIR dependencies, store the element paths from the
snapshot. The paths are stored with the existing entity information.
Then, when looking up element paths, traverse through local definitions
based on "Parent" or "InstanceOf" keyword values until a value that is
not present in the local definitions is reached. This value is hopefully
present in the dependency information. If the value is null, it may be
possible to use the default parent for the entity type.

* Provide path completion items

Any element path that could come after the existing path (which may be
empty) is a potential completion item when writing a rule.

* Store additional element path information

When storing the element information for a FHIR entity, store the path
name and types. Additionally, use a tree structure to store path
information, rather than keeping all element paths as a flat list. When
reading in predefined resources, when a profile is found, rather than
reading its elements, use its type field in order to store the entity
being profiled. Then, refer to that entity when searching for paths for
that profile.

* Create ElementInfo for each type allowed by a choice element

When a choice element is found in a FHIR entity, add the choice
element's info as well as an ElementInfo for each type available on the
choice element. These elements are siblings. Note that if the type of
the choice element is constrained by an OnlyRule, the extension does not
detect this, and will continue to offer the excluded types as completion
items.

* Expand leaf elements to provide completion items

When attempting to complete a path, if the path defines an element that
has no known children but has a known type, use the type to create the
children of that element. When creating children for an element with
more than one type, only include new elements if they are compatible
with all types.

Fix bug in FshCompletionProvider.getElementPathInformation where the
name of a choice element would be incomplete due to overzealous slice
name removal.

* Remove unused function

* Trigger path completion on . character

* Do not provide path completion after the path part of a rule

If the cursor is past a point where there are non-space characters
followed by a space character, this is not the part of the rule where
the path should appear.

* Support URL references for completion items

When creating completion items from dependency definitions, create an
element for the URL if it is available. When creating completion items
from an Alias, store the value of the Alias. Together, this makes it
possible to use an Alias as a parent and get the correct elements for
rule path completion.

Make most properties in NameInfo optional, since many of them are
applicable only to specific entity types.

* Update readme with path completion

The documentation explains the limitations of the element path
autocomplete feature. This now includes not providing completion
elements for indented rules.

* Revise getPathItems for clarity

The new implementation of FshCompletionProvider.getPathItems uses only
one variable to track the traversal through the element tree.

Since the baseDefinition property of EnhancedCompletionItem is populated
by the JSON definition's type property (as opposed to the JSON
definition's baseDefinition property), the property is also named type.

Add some clarifying comments to various functions in
FshCompletionProvider to help explain some of the more unusual steps.

* README update, because Instances get completion items too