FHIR documentation opens page corresponding to project's FHIR version
FshCompletionProvider stores the FHIR version found in the project's
configuration. When opening FHIR documentation, use that FHIR version to
open the documentation page corresponding to the configured version.
Only versions 4.0.1 and later have their corresponding documentation
pages available, as these are the FHIR versions supported by SUSHI. If
the FHIR version is unrecognized or unavailable, open the default
documentation page for the specified name.

Update a handful of dependencies that npm audit was complaining about.